Hook up UI to initial REST endpoints
- Hook up getOntologies, getEntities, and getAxioms
- Disable unnecessary mock endpoints
- Fix a bug in the server.sh script

+++ server.sh
@@ -29,5 +29,8 @@ while (( $serving )); do
 	mvn package exec:java &
 	pid=$!
 	fswatch -1 src/main
-	kill $pid
+	ps -p $pid &> /dev/null
+	if (( $? == 0 )); then
+		kill $pid
+	fi
 done
